added field userId field to user model

+++ controllers/user.js
@@ -1,3 +1,5 @@
+const _ = require("lodash");
+
 module.exports = (sequelize, DataTypes) => {
   const User = sequelize.import("../models/user");
 
@@ -7,7 +9,8 @@ module.exports = (sequelize, DataTypes) => {
     // Returns a promise
     create: params => {
       return User.create({
-        username: params.username,
+        userId: params.userId,
+        userName: params.userName,
         name: params.name,
         bio: params.bio,
         dob: Date.parse(params.dob),
@@ -23,13 +26,13 @@ module.exports = (sequelize, DataTypes) => {
 
     update: params => {
       return User.update(params, {
-        where: { username: params.username },
+        where: { userId: params.userId },
         returning: true
       });
     },
 
     delete: params => {
-      return User.destroy({ where: params });
+      return User.destroy({ where: { userId: params.userId } });
     },
 
     // Deletes all the records from 'users' collection. Does not delete the table
